+++ app/(windows)/us/page.tsx
@@ -15,7 +15,7 @@ export const metadata: Metadata = {
 
 export default function UsPage() {
   return (
-    <div className="grid gap-4 grid-cols-1 md:grid-cols-2 lg:grid-cols-3">
+    <div className="grid gap-5 grid-cols-1 md:grid-cols-2 lg:grid-cols-3">
       <GlobalIndicesCard />
       <Magnificent7Card />
       <UsMoversCard />