+++ components/cards/KrCards.tsx
@@ -18,7 +18,7 @@ import { LongtermDisclosureCard } from "./LongtermCards";
  */
 export function KrCards() {
   return (
-    <div className="grid gap-4 grid-cols-1 md:grid-cols-2 lg:grid-cols-3">
+    <div className="grid gap-5 grid-cols-1 md:grid-cols-2 lg:grid-cols-3">
       <MoversCard />
       <VolumeCard />
       <NetBuyBrokerCard />